requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.3
asttokens==2.1.0
async-timeout==3.0.1
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
decorator==5.1.1
discord.py==1.7.3
executing==1.2.0
idna==3.4
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
stack-data==0.6.0
toml==0.10.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
yarl==1.8.1
